requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cs50==9.4.0
Flask==3.1.0
Flask-Session==0.8.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.39
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
